UPDATE: COURSES LIST ADDED AND PROFILE IMAGE BUG FIXED (#183)
* BUG FIX: NULL CHECK ADDED IN PROFILE AND TEAM IMAGE CHECK FOR NEW USER

* ESLINT CHECKS FIX: FORMATTING FIXED IN ALL FILES

* FEATURE UPDATE: COURSES ARE NOW SENT SEPARATELY FOR EACH USER-ASSIGNED ROADMAP TO BE USED ON learning-path

* BUG FIX: UPDATE USER WAS SHARED SERVICE ADDED BOOLEAN CHECK imageDeleteRequired TO CHECK IF IMAGE IS REQUIRED TO DELETE

* ESLINT FORMATTING: FIXED IN ALL FILES

* DEFAULT PARAM ADDED FOR imageDeleteReqired

+++ apps/quick-learn-backend/src/routes/users/users.service.ts
@@ -240,7 +240,11 @@ export class UsersService extends PaginationService<UserEntity> {
     });
   }
 
-  async updateUser(uuid: UserEntity['uuid'], payload: Partial<UserEntity>) {
+  async updateUser(
+    uuid: UserEntity['uuid'],
+    payload: Partial<UserEntity>,
+    imageDeleteRequired = false,
+  ) {
     const user = await this.findOne({ uuid });
 
     if (!user) {
@@ -249,8 +253,10 @@ export class UsersService extends PaginationService<UserEntity> {
 
     // ON PROFILE CHANGE VERIFY IF LOGO HAS CHANGED AND PERVIOUS IMAGE IS NOT EMPTY STRING
     if (
+      imageDeleteRequired &&
       user.profile_image !== payload.profile_image &&
-      user.profile_image !== ''
+      user.profile_image !== '' &&
+      user.profile_image !== null
     ) {
       // DELETE OLD IMAGE FROM S3 BUCKET
       await this.FileService.deleteFiles([user.profile_image]);
